Feat(test-common): Create test-common sub-project and begin moving common test utilities there
