+++ settings.gradle
@@ -10,4 +10,5 @@ include 'ui'
 include 'web'
 include 'workflow'
 include 'descriptor-api'
+include 'test-common'
