--- src/test/java/com/synopsys/integration/alert/util/TestTags.java
@@ -1,11 +1,0 @@
-package com.synopsys.integration.alert.util;
-
-public class TestTags {
-    // Provided by the common-gradle-plugin
-    public static final String DEFAULT_INTEGRATION = "integration";
-    public static final String DEFAULT_PERFORMANCE = "performance";
-    // Defined by Alert's build.gradle
-    public static final String CUSTOM_BLACKDUCK_CONNECTION = "BlackDuckConnection";
-    public static final String CUSTOM_DATABASE_CONNECTION = "DatabaseConnection";
-    public static final String CUSTOM_EXTERNAL_CONNECTION = "ExternalConnection";
-}
