+++ test-common/build.gradle
@@ -1,0 +1,26 @@
+plugins {
+    id 'java'
+}
+
+group 'com.synopsys.integration'
+version '6.4.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation platform(project(':alert-platform'))
+
+    implementation 'org.springframework:spring-test'
+    implementation 'org.springframework.security:spring-security-test'
+    implementation 'org.springframework.boot:spring-boot-test-autoconfigure'
+    implementation 'org.junit.jupiter:junit-jupiter-params'
+    implementation 'com.github.springtestdbunit:spring-test-dbunit'
+    implementation 'org.mockito:mockito-core'
+    implementation 'org.skyscreamer:jsonassert'
+
+    //    implementation 'jakarta.persistence:jakarta.persistence-api'
+    //    implementation 'org.javassist:javassist'
+    //    implementation 'javax.mail:javax.mail-api'
+}
