+++ test-common/src/main/java/com/synopsys/integration/alert/util/TestTags.java
@@ -1,0 +1,34 @@
+/**
+ * test-common
+ *
+ * Copyright (c) 2020 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.synopsys.integration.alert.util;
+
+public class TestTags {
+    // Provided by the common-gradle-plugin
+    public static final String DEFAULT_INTEGRATION = "integration";
+    public static final String DEFAULT_PERFORMANCE = "performance";
+    // Defined by Alert's build.gradle
+    public static final String CUSTOM_BLACKDUCK_CONNECTION = "BlackDuckConnection";
+    public static final String CUSTOM_DATABASE_CONNECTION = "DatabaseConnection";
+    public static final String CUSTOM_EXTERNAL_CONNECTION = "ExternalConnection";
+
+}
